outflow service(paybill), added custom exceptions

+++ src/main/java/com/aldrichjohne/budget/tracker/model/entity/Bills.java
@@ -5,6 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Entity
@@ -17,6 +18,6 @@ public class Bills {
     @GeneratedValue(strategy = GenerationType.AUTO)
     private UUID id;
     private String name;
-    private double budget;
+    private BigDecimal budget;
     private boolean isDeleted;
 }
