+++ src/main/java/com/aldrichjohne/budget/tracker/model/entity/dto/BillsDTO.java
@@ -3,6 +3,7 @@ package com.aldrichjohne.budget.tracker.model.entity.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Data
@@ -10,6 +11,6 @@ import java.util.UUID;
 public class BillsDTO {
     private UUID id;
     private String name;
-    private double budget;
+    private BigDecimal budget;
     private boolean isDeleted;
 }
